+++ libs/pizza-parser/types/parser.ts
@@ -1,5 +1,5 @@
 import { Pizza } from './pizza';
 
-export interface PizzasParser {
-  parsePizzas(): Promise<Pizza[]>;
+export abstract class PizzasParser {
+  public abstract parsePizzas(): Promise<Pizza[]>;
 }