+++ libs/pizza-parser/types/pizza.ts
@@ -1,16 +1,8 @@
-import { Ingredient } from './ingredient';
-
-export interface Variant {
-  size: number; // 30cm
-  price: number; // 135grn
-  weight: number; // 400g
-}
-
 export const supportedLangs = ['uk', 'ru', 'en'] as const;
 export type Lang = typeof supportedLangs[number];
 
 export const supportedCountries = ['ukraine'] as const;
-export type County = typeof supportedCountries[number];
+export type Country = typeof supportedCountries[number];
 
 export const supportedCities = ['chernivtsi'] as const;
 export type City = typeof supportedCities[number];
@@ -20,10 +12,11 @@ export interface Pizza {
   description: string; // with pepper and pepperoni
   link: string; // http://pizza.com/buy-pizza/:id
   image: string; // http://pizza.com/image
-  ingredients: Ingredient[]; // ['peper', 'pepperoni']
   lang: Lang;
-  variants: Variant[];
-  country: County;
+  size: number | null; // 30cm
+  price: number | null; // 135grn
+  weight: number | null; // 400g
+  country: Country;
   city: City;
 }
 
@@ -34,3 +27,7 @@ export type TranslatedPizza = Omit<Pizza, 'title' | 'description'> &
   {
     [key in TranslatedPizzaTitles | TranslatedPizzaDescription]: string;
   };
+
+export type TranslatedPizzaWithId = TranslatedPizza & {
+  id: string;
+};