+++ libs/pizza-parser/ukraine/chernivtsi/appeti/appeti.spec.ts
@@ -1,0 +1,4 @@
+import { pizzasParserTestSuit } from 'libs/pizza-parser/utils/test';
+import { Apetti } from '.';
+
+pizzasParserTestSuit(Apetti);