+++ libs/pizza-parser/ukraine/chernivtsi/chelentano/chelentano.spec.ts
@@ -1,0 +1,4 @@
+import { pizzasParserTestSuit } from 'libs/pizza-parser/utils/test';
+import { Chelentano } from '.';
+
+pizzasParserTestSuit(Chelentano);