+++ libs/pizza-parser/ukraine/chernivtsi/index.ts
@@ -1,5 +1,11 @@
 import { Apetti } from './appeti';
 import { Chelentano } from './chelentano';
+import { Czernowizza } from './czernowizza';
+import { Dongustavo } from './dongustavo';
+import { Lapiec } from './lapiec';
+import { Misteram } from './misteram';
+import { Monopizza } from './monopizza';
+import { PizzaIt } from './pizza-it';
 import { ShoSho } from './shosho';
 
-export default [ShoSho, Chelentano, Apetti];
+export default [ShoSho, Chelentano, Apetti, Monopizza, Dongustavo, Lapiec, Czernowizza, PizzaIt, Misteram];