+++ libs/pizza-parser/ukraine/chernivtsi/lapiec/lapiec.spec.ts
@@ -1,0 +1,4 @@
+import { pizzasParserTestSuit } from 'libs/pizza-parser/utils/test';
+import { Lapiec } from '.';
+
+pizzasParserTestSuit(Lapiec);