+++ libs/pizza-parser/ukraine/chernivtsi/monopizza/monopizza.spec.ts
@@ -1,0 +1,4 @@
+import { pizzasParserTestSuit } from 'libs/pizza-parser/utils/test';
+import { Monopizza } from '.';
+
+pizzasParserTestSuit(Monopizza);