+++ libs/pizza-parser/ukraine/chernivtsi/pizza-it/pizza-it.spec.ts
@@ -1,0 +1,4 @@
+import { pizzasParserTestSuit } from 'libs/pizza-parser/utils/test';
+import { PizzaIt } from '.';
+
+pizzasParserTestSuit(PizzaIt);