+++ src/app.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { FtsModule } from './fts/fts.module';
 import { PizzaModule } from './pizza/pizza.module';
 
 @Module({
-  imports: [ConfigModule.forRoot(), FtsModule, PizzaModule],
+  imports: [ConfigModule.forRoot(), PizzaModule],
 })
 export class AppModule {}