+++ src/pizza/entities/pizza.entity.ts
@@ -1,26 +1,13 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { Ingredient } from 'libs/pizza-parser/types/ingredient';
 import {
   City,
-  County,
+  Country,
   Lang,
   supportedCities,
   supportedCountries,
   supportedLangs,
   TranslatedPizza,
-  Variant as VariantType,
 } from 'libs/pizza-parser/types/pizza';
-
-class Variant implements VariantType {
-  @ApiProperty({ description: 'pizza weight in grams' })
-  weight: number;
-
-  @ApiProperty({ description: 'pizza size in cm' })
-  size: number;
-
-  @ApiProperty({ description: 'pizza price in grn' })
-  price: number;
-}
 
 export class Pizza implements TranslatedPizza {
   @ApiProperty({ description: 'pizza id what changes after new insert' })
@@ -36,16 +23,19 @@ export class Pizza implements TranslatedPizza {
   lang: Lang;
 
   @ApiProperty({ enum: supportedCountries, description: 'country where pizzeria located' })
-  country: County;
+  country: Country;
 
   @ApiProperty({ enum: supportedCities, description: 'city where pizzeria locatd' })
   city: City;
 
-  @ApiProperty({ enum: Ingredient, isArray: true, description: 'pizza ingredient list' })
-  ingredients: Ingredient[];
+  @ApiProperty({ description: 'pizza weight in grams', nullable: true })
+  weight: number | null;
 
-  @ApiProperty({ type: [Variant], description: 'pizza variant list' })
-  variants: VariantType[];
+  @ApiProperty({ description: 'pizza size in cm', nullable: true })
+  size: number | null;
+
+  @ApiProperty({ description: 'pizza price in grn', nullable: true })
+  price: number | null;
 
   @ApiProperty({ description: 'pizza title in ukrainian' })
   uk_title: string;