+++ src/pizza/pizza.module.ts
@@ -1,10 +1,17 @@
 import { Module } from '@nestjs/common';
 import { PizzaService } from './pizza.service';
 import { PizzaController } from './pizza.controller';
-import { FtsModule } from 'src/fts/fts.module';
+import { ElasticsearchModule } from '@nestjs/elasticsearch';
+import { ConfigModule, ConfigService } from '@nestjs/config';
 
 @Module({
-  imports: [FtsModule],
+  imports: [
+    ElasticsearchModule.registerAsync({
+      imports: [ConfigModule],
+      useFactory: (configService: ConfigService) => ({ node: configService.get('ES_URL') }),
+      inject: [ConfigService],
+    }),
+  ],
   controllers: [PizzaController],
   providers: [PizzaService],
 })