+++ src/pizza/pizza.service.ts
@@ -1,70 +1,186 @@
 import fsp from 'fs/promises';
 import path from 'path';
 import { Injectable, OnModuleInit } from '@nestjs/common';
-import { FtsService } from 'src/fts/fts.service';
-import { TranslatedPizza } from 'libs/pizza-parser/types/pizza';
-import { delay } from 'src/utils/delay';
-import { nanoid } from 'nanoid';
+import { TranslatedPizzaWithId } from 'libs/pizza-parser/types/pizza';
 import { GetPizzasDto } from './dto/getPizzas.dto';
-import { FtsWhereBuilder } from 'src/fts/fts-where.builder';
 import { Pizza } from './entities/pizza.entity';
+import { ElasticsearchService } from '@nestjs/elasticsearch';
+import { estypes } from '@elastic/elasticsearch';
+import { SearchQuery } from 'src/types/elasticsearch';
 
 interface SavedPizzas {
   timestamp: number;
-  pizzas: TranslatedPizza[];
+  pizzas: TranslatedPizzaWithId[];
 }
+
 const pizzasPath = path.join(process.cwd(), 'pizzas.json');
+
+const numberAndText = {
+  type: 'integer',
+  fields: {
+    text: {
+      type: 'text',
+    },
+  },
+} as const;
 
 @Injectable()
 export class PizzaService implements OnModuleInit {
-  private pizzasIndex = 'pizzas';
-  private pizzasTimestampIndex = 'pizzas-timestamp';
+  private pizzasIndexBase = 'pizzas';
+  private pizzasIndex: string;
 
-  constructor(private ftsService: FtsService) {}
+  constructor(private elasticsearchService: ElasticsearchService) {}
 
   public async onModuleInit() {
-    const { timestamp, pizzas } = JSON.parse(await fsp.readFile(pizzasPath, 'utf-8')) as SavedPizzas;
-
-    const index = await this.ftsService.client.getOrCreateIndex(this.pizzasTimestampIndex);
-    await index.updateFilterableAttributes(['timestamp']);
-
-    await delay(1000);
-    const { hits } = await index.search('', { filter: `timestamp = ${timestamp}` });
-
-    if (hits.length !== 0) {
-      return;
-    }
-
-    const pizzaIndex = await this.ftsService.client.getOrCreateIndex(this.pizzasIndex);
-
-    await pizzaIndex.deleteAllDocuments();
-    await delay(1000);
-    await pizzaIndex.updateFilterableAttributes(['city', 'country', 'id']);
-    await delay(1000);
-
-    await pizzaIndex.addDocuments(pizzas.map((pizza) => ({ ...pizza, id: nanoid() })));
-    await index.addDocuments([{ timestamp, id: nanoid() }]);
+    await this.setPizzasIfNotExists();
   }
 
-  public async getPizzas({ query, limit, offset, country, city }: GetPizzasDto) {
-    const where = new FtsWhereBuilder().equal('country', country).and().equal('city', city).build();
-    const result = await this.ftsService.select<Pizza>(this.pizzasIndex, query, { limit, offset, where });
+  private async setPizzasIfNotExists() {
+    const { timestamp, pizzas } = JSON.parse(await fsp.readFile(pizzasPath, 'utf-8')) as SavedPizzas;
 
-    return result;
+    this.pizzasIndex = `${this.pizzasIndexBase}-${timestamp}`;
+
+    const { body: isExists } = await this.elasticsearchService.indices.exists({ index: this.pizzasIndex });
+
+    if (isExists) return;
+
+    const mappings: estypes.MappingTypeMapping = {
+      properties: {
+        id: { type: 'text' },
+        image: { type: 'text' },
+        link: { type: 'text' },
+        lang: { type: 'text' },
+        country: { type: 'text' },
+        city: { type: 'text' },
+        weight: numberAndText,
+        size: numberAndText,
+        price: numberAndText,
+        uk_title: { type: 'text' },
+        uk_description: { type: 'text' },
+        ru_title: { type: 'text' },
+        ru_description: { type: 'text' },
+        en_title: { type: 'text' },
+        en_description: { type: 'text' },
+      },
+    };
+
+    await this.elasticsearchService.indices.create<estypes.IndicesCreateResponse>({
+      index: this.pizzasIndex,
+      body: {
+        mappings,
+      },
+    });
+
+    const bulkBody = pizzas.flatMap((pizza) => [{ index: { _index: this.pizzasIndex } }, pizza]);
+    const {
+      body: { errors },
+    } = await this.elasticsearchService.bulk<estypes.BulkResponse>({ refresh: true, body: bulkBody });
+
+    if (errors) throw new Error('bulk has errors');
+  }
+
+  private queryToSimpleQuery(query: string) {
+    return query
+      .split(' ')
+      .map((word) => {
+        const isNumber = !isNaN(parseInt(word));
+
+        if (isNumber) return word;
+        return `${word}~`;
+      })
+      .join(' ');
+  }
+
+  public async getPizzas({ query, limit, offset, country, city, orderBy }: GetPizzasDto) {
+    const must: estypes.QueryDslQueryContainer = {};
+
+    if (query.length === 0) {
+      must.match_all = {};
+    } else {
+      must.simple_query_string = {
+        query: this.queryToSimpleQuery(query),
+        fields: ['*_title', '*_description', 'weight.text', 'size.text', 'price.text'],
+        default_operator: 'and',
+      };
+    }
+
+    const esQuery: SearchQuery = {
+      index: this.pizzasIndex,
+      body: {
+        from: offset,
+        size: limit,
+        sort: orderBy ? [{ [orderBy.target]: orderBy.cause }] : undefined,
+        query: {
+          bool: {
+            must,
+            filter: [{ term: { country } }, { term: { city } }],
+          },
+        },
+      },
+    };
+
+    const {
+      body: { hits },
+    } = await this.elasticsearchService.search<estypes.SearchResponse<Pizza>>(esQuery);
+
+    const total: number = typeof hits.total === 'number' ? hits.total : hits.total.value;
+    const value = hits.hits.map(({ _source }) => _source);
+
+    return { total, value };
   }
 
   public async getPizzasTotal() {
-    return await this.ftsService.totalCount(this.pizzasIndex);
+    const {
+      body: { count: total },
+    } = await this.elasticsearchService.count<estypes.CountResponse>({
+      index: this.pizzasIndex,
+    });
+
+    return { total };
   }
 
   public async getPizzasByIds(ids: string[]) {
-    const value = await this.ftsService.getByIds<Pizza>(this.pizzasIndex, ids);
+    const {
+      body: {
+        hits: { hits },
+      },
+    } = await this.elasticsearchService.search<estypes.SearchResponse<Pizza>>({
+      index: this.pizzasIndex,
+      body: {
+        query: {
+          bool: {
+            should: ids.map((id) => ({ match: { id } })),
+          },
+        },
+      },
+    });
+
+    const value = hits.map(({ _source }) => _source);
 
     return { value };
   }
 
   public async getPizzaById(id: string) {
-    const value = await this.ftsService.getById<Pizza>(this.pizzasIndex, id);
+    const {
+      body: {
+        hits: { hits },
+      },
+    } = await this.elasticsearchService.search<estypes.SearchResponse<Pizza>>({
+      index: this.pizzasIndex,
+      body: {
+        query: {
+          bool: {
+            must: [
+              {
+                match: { id },
+              },
+            ],
+          },
+        },
+      },
+    });
+
+    const value = hits.map(({ _source }) => _source)[0];
 
     return { value };
   }